refactor v2 eval let step tests
